+++ src/Event/ClickCompleted.php
@@ -1,7 +1,7 @@
 <?php
 namespace Athena\Event;
 
-use Athena\Browser\BrowserInterface;
+use OLX\FluentWebDriverClient\Browser\BrowserInterface;
 use Symfony\Component\EventDispatcher\Event;
 
 class ClickCompleted extends Event
@@ -9,14 +9,14 @@ class ClickCompleted extends Event
     const AFTER = 'browser.click.after';
 
     /**
-     * @var \Athena\Browser\BrowserInterface
+     * @var \OLX\FluentWebDriverClient\Browser\BrowserInterface
      */
     private $browser;
 
     /**
      * ClickCompleted constructor.
      *
-     * @param \Athena\Browser\BrowserInterface $browser
+     * @param \OLX\FluentWebDriverClient\Browser\BrowserInterface $browser
      */
     public function __construct(BrowserInterface $browser)
     {