+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.7.2 (December 12, 2016)
+
+- Restored `BrowserDriverBuilder::fromSettings`
+
+## 0.7.1 (December 8, 2016)
+
+- Added `composer` sub-command
+- FluentWebDriverClient is now an external dependency
+- Documentation updates
+- Updated FluentHttpClient dependency
+- Allow setting a default value for ENV variables used in configuration
+
 ## 0.7.0 (September 9, 2016)
 
 ### Athena Plugin PHP